Tide Date Times

Generates sequence of date times.

tide_datetimes(minutes = 60L, from = as.Date("2015-01-01"),
  to = as.Date("2015-12-31"), tz = "PST8PDT")

Arguments

minutes
An integer of the number of minutes between tide heights

from
A Date of the start of the period of interest

to
A Date of the end of the period of interest

tz
A string of the time zone.

Value

A POSIXct vector.

Examples

tide_datetimes()

#>    [1] "2015-01-01 00:00:00 PST" "2015-01-01 01:00:00 PST"
#>    [3] "2015-01-01 02:00:00 PST" "2015-01-01 03:00:00 PST"
#>    [5] "2015-01-01 04:00:00 PST" "2015-01-01 05:00:00 PST"
#>    [7] "2015-01-01 06:00:00 PST" "2015-01-01 07:00:00 PST"
#>    [9] "2015-01-01 08:00:00 PST" "2015-01-01 09:00:00 PST"
#>   [11] "2015-01-01 10:00:00 PST" "2015-01-01 11:00:00 PST"
#>   [13] "2015-01-01 12:00:00 PST" "2015-01-01 13:00:00 PST"
#>   [15] "2015-01-01 14:00:00 PST" "2015-01-01 15:00:00 PST"
#>   [17] "2015-01-01 16:00:00 PST" "2015-01-01 17:00:00 PST"
#>   [19] "2015-01-01 18:00:00 PST" "2015-01-01 19:00:00 PST"
#>   [21] "2015-01-01 20:00:00 PST" "2015-01-01 21:00:00 PST"
#>   [23] "2015-01-01 22:00:00 PST" "2015-01-01 23:00:00 PST"
#>   [25] "2015-01-02 00:00:00 PST" "2015-01-02 01:00:00 PST"
#>   [27] "2015-01-02 02:00:00 PST" "2015-01-02 03:00:00 PST"
#>   [29] "2015-01-02 04:00:00 PST" "2015-01-02 05:00:00 PST"
#>   [31] "2015-01-02 06:00:00 PST" "2015-01-02 07:00:00 PST"
#>   [33] "2015-01-02 08:00:00 PST" "2015-01-02 09:00:00 PST"
#>   [35] "2015-01-02 10:00:00 PST" "2015-01-02 11:00:00 PST"
#>   [37] "2015-01-02 12:00:00 PST" "2015-01-02 13:00:00 PST"
#>   [39] "2015-01-02 14:00:00 PST" "2015-01-02 15:00:00 PST"
#>   [41] "2015-01-02 16:00:00 PST" "2015-01-02 17:00:00 PST"
#>   [43] "2015-01-02 18:00:00 PST" "2015-01-02 19:00:00 PST"
#>   [45] "2015-01-02 20:00:00 PST" "2015-01-02 21:00:00 PST"
#>   [47] "2015-01-02 22:00:00 PST" "2015-01-02 23:00:00 PST"
#>   [49] "2015-01-03 00:00:00 PST" "2015-01-03 01:00:00 PST"
#>   [51] "2015-01-03 02:00:00 PST" "2015-01-03 03:00:00 PST"
#>   [53] "2015-01-03 04:00:00 PST" "2015-01-03 05:00:00 PST"
#>   [55] "2015-01-03 06:00:00 PST" "2015-01-03 07:00:00 PST"
#>   [57] "2015-01-03 08:00:00 PST" "2015-01-03 09:00:00 PST"
#>   [59] "2015-01-03 10:00:00 PST" "2015-01-03 11:00:00 PST"
#>   [61] "2015-01-03 12:00:00 PST" "2015-01-03 13:00:00 PST"
#>   [63] "2015-01-03 14:00:00 PST" "2015-01-03 15:00:00 PST"
#>   [65] "2015-01-03 16:00:00 PST" "2015-01-03 17:00:00 PST"
#>   [67] "2015-01-03 18:00:00 PST" "2015-01-03 19:00:00 PST"
#>   [69] "2015-01-03 20:00:00 PST" "2015-01-03 21:00:00 PST"
#>   [71] "2015-01-03 22:00:00 PST" "2015-01-03 23:00:00 PST"
#>   [73] "2015-01-04 00:00:00 PST" "2015-01-04 01:00:00 PST"
#>   [75] "2015-01-04 02:00:00 PST" "2015-01-04 03:00:00 PST"
#>   [77] "2015-01-04 04:00:00 PST" "2015-01-04 05:00:00 PST"
#>   [79] "2015-01-04 06:00:00 PST" "2015-01-04 07:00:00 PST"
#>   [81] "2015-01-04 08:00:00 PST" "2015-01-04 09:00:00 PST"
#>   [83] "2015-01-04 10:00:00 PST" "2015-01-04 11:00:00 PST"
#>   [85] "2015-01-04 12:00:00 PST" "2015-01-04 13:00:00 PST"
#>   [87] "2015-01-04 14:00:00 PST" "2015-01-04 15:00:00 PST"
#>   [89] "2015-01-04 16:00:00 PST" "2015-01-04 17:00:00 PST"
#>   [91] "2015-01-04 18:00:00 PST" "2015-01-04 19:00:00 PST"
#>   [93] "2015-01-04 20:00:00 PST" "2015-01-04 21:00:00 PST"
#>   [95] "2015-01-04 22:00:00 PST" "2015-01-04 23:00:00 PST"
#>   [97] "2015-01-05 00:00:00 PST" "2015-01-05 01:00:00 PST"
#>   [99] "2015-01-05 02:00:00 PST" "2015-01-05 03:00:00 PST"
#>  [101] "2015-01-05 04:00:00 PST" "2015-01-05 05:00:00 PST"
#>  [103] "2015-01-05 06:00:00 PST" "2015-01-05 07:00:00 PST"
#>  [105] "2015-01-05 08:00:00 PST" "2015-01-05 09:00:00 PST"
#>  [107] "2015-01-05 10:00:00 PST" "2015-01-05 11:00:00 PST"
#>  [109] "2015-01-05 12:00:00 PST" "2015-01-05 13:00:00 PST"
#>  [111] "2015-01-05 14:00:00 PST" "2015-01-05 15:00:00 PST"
#>  [113] "2015-01-05 16:00:00 PST" "2015-01-05 17:00:00 PST"
#>  [115] "2015-01-05 18:00:00 PST" "2015-01-05 19:00:00 PST"
#>  [117] "2015-01-05 20:00:00 PST" "2015-01-05 21:00:00 PST"
#>  [119] "2015-01-05 22:00:00 PST" "2015-01-05 23:00:00 PST"
#>  [121] "2015-01-06 00:00:00 PST" "2015-01-06 01:00:00 PST"
#>  [123] "2015-01-06 02:00:00 PST" "2015-01-06 03:00:00 PST"
#>  [125] "2015-01-06 04:00:00 PST" "2015-01-06 05:00:00 PST"
#>  [127] "2015-01-06 06:00:00 PST" "2015-01-06 07:00:00 PST"
#>  [129] "2015-01-06 08:00:00 PST" "2015-01-06 09:00:00 PST"
#>  [131] "2015-01-06 10:00:00 PST" "2015-01-06 11:00:00 PST"
#>  [133] "2015-01-06 12:00:00 PST" "2015-01-06 13:00:00 PST"
#>  [135] "2015-01-06 14:00:00 PST" "2015-01-06 15:00:00 PST"
#>  [137] "2015-01-06 16:00:00 PST" "2015-01-06 17:00:00 PST"
#>  [139] "2015-01-06 18:00:00 PST" "2015-01-06 19:00:00 PST"
#>  [141] "2015-01-06 20:00:00 PST" "2015-01-06 21:00:00 PST"
#>  [143] "2015-01-06 22:00:00 PST" "2015-01-06 23:00:00 PST"
#>  [145] "2015-01-07 00:00:00 PST" "2015-01-07 01:00:00 PST"
#>  [147] "2015-01-07 02:00:00 PST" "2015-01-07 03:00:00 PST"
#>  [149] "2015-01-07 04:00:00 PST" "2015-01-07 05:00:00 PST"
#>  [151] "2015-01-07 06:00:00 PST" "2015-01-07 07:00:00 PST"
#>  [153] "2015-01-07 08:00:00 PST" "2015-01-07 09:00:00 PST"
#>  [155] "2015-01-07 10:00:00 PST" "2015-01-07 11:00:00 PST"
#>  [157] "2015-01-07 12:00:00 PST" "2015-01-07 13:00:00 PST"
#>  [159] "2015-01-07 14:00:00 PST" "2015-01-07 15:00:00 PST"
#>  [161] "2015-01-07 16:00:00 PST" "2015-01-07 17:00:00 PST"
#>  [163] "2015-01-07 18:00:00 PST" "2015-01-07 19:00:00 PST"
#>  [165] "2015-01-07 20:00:00 PST" "2015-01-07 21:00:00 PST"
#>  [167] "2015-01-07 22:00:00 PST" "2015-01-07 23:00:00 PST"
#>  [169] "2015-01-08 00:00:00 PST" "2015-01-08 01:00:00 PST"
#>  [171] "2015-01-08 02:00:00 PST" "2015-01-08 03:00:00 PST"
#>  [173] "2015-01-08 04:00:00 PST" "2015-01-08 05:00:00 PST"
#>  [175] "2015-01-08 06:00:00 PST" "2015-01-08 07:00:00 PST"
#>  [177] "2015-01-08 08:00:00 PST" "2015-01-08 09:00:00 PST"
#>  [179] "2015-01-08 10:00:00 PST" "2015-01-08 11:00:00 PST"
#>  [181] "2015-01-08 12:00:00 PST" "2015-01-08 13:00:00 PST"
#>  [183] "2015-01-08 14:00:00 PST" "2015-01-08 15:00:00 PST"
#>  [185] "2015-01-08 16:00:00 PST" "2015-01-08 17:00:00 PST"
#>  [187] "2015-01-08 18:00:00 PST" "2015-01-08 19:00:00 PST"
#>  [189] "2015-01-08 20:00:00 PST" "2015-01-08 21:00:00 PST"
#>  [191] "2015-01-08 22:00:00 PST" "2015-01-08 23:00:00 PST"
#>  [193] "2015-01-09 00:00:00 PST" "2015-01-09 01:00:00 PST"
#>  [195] "2015-01-09 02:00:00 PST" "2015-01-09 03:00:00 PST"
#>  [197] "2015-01-09 04:00:00 PST" "2015-01-09 05:00:00 PST"
#>  [199] "2015-01-09 06:00:00 PST" "2015-01-09 07:00:00 PST"
#>  [201] "2015-01-09 08:00:00 PST" "2015-01-09 09:00:00 PST"
#>  [203] "2015-01-09 10:00:00 PST" "2015-01-09 11:00:00 PST"
#>  [205] "2015-01-09 12:00:00 PST" "2015-01-09 13:00:00 PST"
#>  [207] "2015-01-09 14:00:00 PST" "2015-01-09 15:00:00 PST"
#>  [209] "2015-01-09 16:00:00 PST" "2015-01-09 17:00:00 PST"
#>  [211] "2015-01-09 18:00:00 PST" "2015-01-09 19:00:00 PST"
#>  [213] "2015-01-09 20:00:00 PST" "2015-01-09 21:00:00 PST"
#>  [215] "2015-01-09 22:00:00 PST" "2015-01-09 23:00:00 PST"
#>  [217] "2015-01-10 00:00:00 PST" "2015-01-10 01:00:00 PST"
#>  [219] "2015-01-10 02:00:00 PST" "2015-01-10 03:00:00 PST"
#>  [221] "2015-01-10 04:00:00 PST" "2015-01-10 05:00:00 PST"
#>  [223] "2015-01-10 06:00:00 PST" "2015-01-10 07:00:00 PST"
#>  [225] "2015-01-10 08:00:00 PST" "2015-01-10 09:00:00 PST"
#>  [227] "2015-01-10 10:00:00 PST" "2015-01-10 11:00:00 PST"
#>  [229] "2015-01-10 12:00:00 PST" "2015-01-10 13:00:00 PST"
#>  [231] "2015-01-10 14:00:00 PST" "2015-01-10 15:00:00 PST"
#>  [233] "2015-01-10 16:00:00 PST" "2015-01-10 17:00:00 PST"
#>  [235] "2015-01-10 18:00:00 PST" "2015-01-10 19:00:00 PST"
#>  [237] "2015-01-10 20:00:00 PST" "2015-01-10 21:00:00 PST"
#>  [239] "2015-01-10 22:00:00 PST" "2015-01-10 23:00:00 PST"
#>  [241] "2015-01-11 00:00:00 PST" "2015-01-11 01:00:00 PST"
#>  [243] "2015-01-11 02:00:00 PST" "2015-01-11 03:00:00 PST"
#>  [245] "2015-01-11 04:00:00 PST" "2015-01-11 05:00:00 PST"
#>  [247] "2015-01-11 06:00:00 PST" "2015-01-11 07:00:00 PST"
#>  [249] "2015-01-11 08:00:00 PST" "2015-01-11 09:00:00 PST"
#>  [251] "2015-01-11 10:00:00 PST" "2015-01-11 11:00:00 PST"
#>  [253] "2015-01-11 12:00:00 PST" "2015-01-11 13:00:00 PST"
#>  [255] "2015-01-11 14:00:00 PST" "2015-01-11 15:00:00 PST"
#>  [257] "2015-01-11 16:00:00 PST" "2015-01-11 17:00:00 PST"
#>  [259] "2015-01-11 18:00:00 PST" "2015-01-11 19:00:00 PST"
#>  [261] "2015-01-11 20:00:00 PST" "2015-01-11 21:00:00 PST"
#>  [263] "2015-01-11 22:00:00 PST" "2015-01-11 23:00:00 PST"
#>  [265] "2015-01-12 00:00:00 PST" "2015-01-12 01:00:00 PST"
#>  [267] "2015-01-12 02:00:00 PST" "2015-01-12 03:00:00 PST"
#>  [269] "2015-01-12 04:00:00 PST" "2015-01-12 05:00:00 PST"
#>  [271] "2015-01-12 06:00:00 PST" "2015-01-12 07:00:00 PST"
#>  [273] "2015-01-12 08:00:00 PST" "2015-01-12 09:00:00 PST"
#>  [275] "2015-01-12 10:00:00 PST" "2015-01-12 11:00:00 PST"
#>  [277] "2015-01-12 12:00:00 PST" "2015-01-12 13:00:00 PST"
#>  [279] "2015-01-12 14:00:00 PST" "2015-01-12 15:00:00 PST"
#>  [281] "2015-01-12 16:00:00 PST" "2015-01-12 17:00:00 PST"
#>  [283] "2015-01-12 18:00:00 PST" "2015-01-12 19:00:00 PST"
#>  [285] "2015-01-12 20:00:00 PST" "2015-01-12 21:00:00 PST"
#>  [287] "2015-01-12 22:00:00 PST" "2015-01-12 23:00:00 PST"
#>  [289] "2015-01-13 00:00:00 PST" "2015-01-13 01:00:00 PST"
#>  [291] "2015-01-13 02:00:00 PST" "2015-01-13 03:00:00 PST"
#>  [293] "2015-01-13 04:00:00 PST" "2015-01-13 05:00:00 PST"
#>  [295] "2015-01-13 06:00:00 PST" "2015-01-13 07:00:00 PST"
#>  [297] "2015-01-13 08:00:00 PST" "2015-01-13 09:00:00 PST"
#>  [299] "2015-01-13 10:00:00 PST" "2015-01-13 11:00:00 PST"
#>  [301] "2015-01-13 12:00:00 PST" "2015-01-13 13:00:00 PST"
#>  [303] "2015-01-13 14:00:00 PST" "2015-01-13 15:00:00 PST"
#>  [305] "2015-01-13 16:00:00 PST" "2015-01-13 17:00:00 PST"
#>  [307] "2015-01-13 18:00:00 PST" "2015-01-13 19:00:00 PST"
#>  [309] "2015-01-13 20:00:00 PST" "2015-01-13 21:00:00 PST"
#>  [311] "2015-01-13 22:00:00 PST" "2015-01-13 23:00:00 PST"
#>  [313] "2015-01-14 00:00:00 PST" "2015-01-14 01:00:00 PST"
#>  [315] "2015-01-14 02:00:00 PST" "2015-01-14 03:00:00 PST"
#>  [317] "2015-01-14 04:00:00 PST" "2015-01-14 05:00:00 PST"
#>  [319] "2015-01-14 06:00:00 PST" "2015-01-14 07:00:00 PST"
#>  [321] "2015-01-14 08:00:00 PST" "2015-01-14 09:00:00 PST"
#>  [323] "2015-01-14 10:00:00 PST" "2015-01-14 11:00:00 PST"
#>  [325] "2015-01-14 12:00:00 PST" "2015-01-14 13:00:00 PST"
#>  [327] "2015-01-14 14:00:00 PST" "2015-01-14 15:00:00 PST"
#>  [329] "2015-01-14 16:00:00 PST" "2015-01-14 17:00:00 PST"
#>  [331] "2015-01-14 18:00:00 PST" "2015-01-14 19:00:00 PST"
#>  [333] "2015-01-14 20:00:00 PST" "2015-01-14 21:00:00 PST"
#>  [335] "2015-01-14 22:00:00 PST" "2015-01-14 23:00:00 PST"
#>  [337] "2015-01-15 00:00:00 PST" "2015-01-15 01:00:00 PST"
#>  [339] "2015-01-15 02:00:00 PST" "2015-01-15 03:00:00 PST"
#>  [341] "2015-01-15 04:00:00 PST" "2015-01-15 05:00:00 PST"
#>  [343] "2015-01-15 06:00:00 PST" "2015-01-15 07:00:00 PST"
#>  [345] "2015-01-15 08:00:00 PST" "2015-01-15 09:00:00 PST"
#>  [347] "2015-01-15 10:00:00 PST" "2015-01-15 11:00:00 PST"
#>  [349] "2015-01-15 12:00:00 PST" "2015-01-15 13:00:00 PST"
#>  [351] "2015-01-15 14:00:00 PST" "2015-01-15 15:00:00 PST"
#>  [353] "2015-01-15 16:00:00 PST" "2015-01-15 17:00:00 PST"
#>  [355] "2015-01-15 18:00:00 PST" "2015-01-15 19:00:00 PST"
#>  [357] "2015-01-15 20:00:00 PST" "2015-01-15 21:00:00 PST"
#>  [359] "2015-01-15 22:00:00 PST" "2015-01-15 23:00:00 PST"
#>  [361] "2015-01-16 00:00:00 PST" "2015-01-16 01:00:00 PST"
#>  [363] "2015-01-16 02:00:00 PST" "2015-01-16 03:00:00 PST"
#>  [365] "2015-01-16 04:00:00 PST" "2015-01-16 05:00:00 PST"
#>  [367] "2015-01-16 06:00:00 PST" "2015-01-16 07:00:00 PST"
#>  [369] "2015-01-16 08:00:00 PST" "2015-01-16 09:00:00 PST"
#>  [371] "2015-01-16 10:00:00 PST" "2015-01-16 11:00:00 PST"
#>  [373] "2015-01-16 12:00:00 PST" "2015-01-16 13:00:00 PST"
#>  [375] "2015-01-16 14:00:00 PST" "2015-01-16 15:00:00 PST"
#>  [377] "2015-01-16 16:00:00 PST" "2015-01-16 17:00:00 PST"
#>  [379] "2015-01-16 18:00:00 PST" "2015-01-16 19:00:00 PST"
#>  [381] "2015-01-16 20:00:00 PST" "2015-01-16 21:00:00 PST"
#>  [383] "2015-01-16 22:00:00 PST" "2015-01-16 23:00:00 PST"
#>  [385] "2015-01-17 00:00:00 PST" "2015-01-17 01:00:00 PST"
#>  [387] "2015-01-17 02:00:00 PST" "2015-01-17 03:00:00 PST"
#>  [389] "2015-01-17 04:00:00 PST" "2015-01-17 05:00:00 PST"
#>  [391] "2015-01-17 06:00:00 PST" "2015-01-17 07:00:00 PST"
#>  [393] "2015-01-17 08:00:00 PST" "2015-01-17 09:00:00 PST"
#>  [395] "2015-01-17 10:00:00 PST" "2015-01-17 11:00:00 PST"
#>  [397] "2015-01-17 12:00:00 PST" "2015-01-17 13:00:00 PST"
#>  [399] "2015-01-17 14:00:00 PST" "2015-01-17 15:00:00 PST"
#>  [401] "2015-01-17 16:00:00 PST" "2015-01-17 17:00:00 PST"
#>  [403] "2015-01-17 18:00:00 PST" "2015-01-17 19:00:00 PST"
#>  [405] "2015-01-17 20:00:00 PST" "2015-01-17 21:00:00 PST"
#>  [407] "2015-01-17 22:00:00 PST" "2015-01-17 23:00:00 PST"
#>  [409] "2015-01-18 00:00:00 PST" "2015-01-18 01:00:00 PST"
#>  [411] "2015-01-18 02:00:00 PST" "2015-01-18 03:00:00 PST"
#>  [413] "2015-01-18 04:00:00 PST" "2015-01-18 05:00:00 PST"
#>  [415] "2015-01-18 06:00:00 PST" "2015-01-18 07:00:00 PST"
#>  [417] "2015-01-18 08:00:00 PST" "2015-01-18 09:00:00 PST"
#>  [419] "2015-01-18 10:00:00 PST" "2015-01-18 11:00:00 PST"
#>  [421] "2015-01-18 12:00:00 PST" "2015-01-18 13:00:00 PST"
#>  [423] "2015-01-18 14:00:00 PST" "2015-01-18 15:00:00 PST"
#>  [425] "2015-01-18 16:00:00 PST" "2015-01-18 17:00:00 PST"
#>  [427] "2015-01-18 18:00:00 PST" "2015-01-18 19:00:00 PST"
#>  [429] "2015-01-18 20:00:00 PST" "2015-01-18 21:00:00 PST"
#>  [431] "2015-01-18 22:00:00 PST" "2015-01-18 23:00:00 PST"
#>  [433] "2015-01-19 00:00:00 PST" "2015-01-19 01:00:00 PST"
#>  [435] "2015-01-19 02:00:00 PST" "2015-01-19 03:00:00 PST"
#>  [437] "2015-01-19 04:00:00 PST" "2015-01-19 05:00:00 PST"
#>  [439] "2015-01-19 06:00:00 PST" "2015-01-19 07:00:00 PST"
#>  [441] "2015-01-19 08:00:00 PST" "2015-01-19 09:00:00 PST"
#>  [443] "2015-01-19 10:00:00 PST" "2015-01-19 11:00:00 PST"
#>  [445] "2015-01-19 12:00:00 PST" "2015-01-19 13:00:00 PST"
#>  [447] "2015-01-19 14:00:00 PST" "2015-01-19 15:00:00 PST"
#>  [449] "2015-01-19 16:00:00 PST" "2015-01-19 17:00:00 PST"
#>  [451] "2015-01-19 18:00:00 PST" "2015-01-19 19:00:00 PST"
#>  [453] "2015-01-19 20:00:00 PST" "2015-01-19 21:00:00 PST"
#>  [455] "2015-01-19 22:00:00 PST" "2015-01-19 23:00:00 PST"
#>  [457] "2015-01-20 00:00:00 PST" "2015-01-20 01:00:00 PST"
#>  [459] "2015-01-20 02:00:00 PST" "2015-01-20 03:00:00 PST"
#>  [461] "2015-01-20 04:00:00 PST" "2015-01-20 05:00:00 PST"
#>  [463] "2015-01-20 06:00:00 PST" "2015-01-20 07:00:00 PST"
#>  [465] "2015-01-20 08:00:00 PST" "2015-01-20 09:00:00 PST"
#>  [467] "2015-01-20 10:00:00 PST" "2015-01-20 11:00:00 PST"
#>  [469] "2015-01-20 12:00:00 PST" "2015-01-20 13:00:00 PST"
#>  [471] "2015-01-20 14:00:00 PST" "2015-01-20 15:00:00 PST"
#>  [473] "2015-01-20 16:00:00 PST" "2015-01-20 17:00:00 PST"
#>  [475] "2015-01-20 18:00:00 PST" "2015-01-20 19:00:00 PST"
#>  [477] "2015-01-20 20:00:00 PST" "2015-01-20 21:00:00 PST"
#>  [479] "2015-01-20 22:00:00 PST" "2015-01-20 23:00:00 PST"
#>  [481] "2015-01-21 00:00:00 PST" "2015-01-21 01:00:00 PST"
#>  [483] "2015-01-21 02:00:00 PST" "2015-01-21 03:00:00 PST"
#>  [485] "2015-01-21 04:00:00 PST" "2015-01-21 05:00:00 PST"
#>  [487] "2015-01-21 06:00:00 PST" "2015-01-21 07:00:00 PST"
#>  [489] "2015-01-21 08:00:00 PST" "2015-01-21 09:00:00 PST"
#>  [491] "2015-01-21 10:00:00 PST" "2015-01-21 11:00:00 PST"
#>  [493] "2015-01-21 12:00:00 PST" "2015-01-21 13:00:00 PST"
#>  [495] "2015-01-21 14:00:00 PST" "2015-01-21 15:00:00 PST"
#>  [497] "2015-01-21 16:00:00 PST" "2015-01-21 17:00:00 PST"
#>  [499] "2015-01-21 18:00:00 PST" "2015-01-21 19:00:00 PST"
#>  [501] "2015-01-21 20:00:00 PST" "2015-01-21 21:00:00 PST"
#>  [503] "2015-01-21 22:00:00 PST" "2015-01-21 23:00:00 PST"
#>  [505] "2015-01-22 00:00:00 PST" "2015-01-22 01:00:00 PST"
#>  [507] "2015-01-22 02:00:00 PST" "2015-01-22 03:00:00 PST"
#>  [509] "2015-01-22 04:00:00 PST" "2015-01-22 05:00:00 PST"
#>  [511] "2015-01-22 06:00:00 PST" "2015-01-22 07:00:00 PST"
#>  [513] "2015-01-22 08:00:00 PST" "2015-01-22 09:00:00 PST"
#>  [515] "2015-01-22 10:00:00 PST" "2015-01-22 11:00:00 PST"
#>  [517] "2015-01-22 12:00:00 PST" "2015-01-22 13:00:00 PST"
#>  [519] "2015-01-22 14:00:00 PST" "2015-01-22 15:00:00 PST"
#>  [521] "2015-01-22 16:00:00 PST" "2015-01-22 17:00:00 PST"
#>  [523] "2015-01-22 18:00:00 PST" "2015-01-22 19:00:00 PST"
#>  [525] "2015-01-22 20:00:00 PST" "2015-01-22 21:00:00 PST"
#>  [527] "2015-01-22 22:00:00 PST" "2015-01-22 23:00:00 PST"
#>  [529] "2015-01-23 00:00:00 PST" "2015-01-23 01:00:00 PST"
#>  [531] "2015-01-23 02:00:00 PST" "2015-01-23 03:00:00 PST"
#>  [533] "2015-01-23 04:00:00 PST" "2015-01-23 05:00:00 PST"
#>  [535] "2015-01-23 06:00:00 PST" "2015-01-23 07:00:00 PST"
#>  [537] "2015-01-23 08:00:00 PST" "2015-01-23 09:00:00 PST"
#>  [539] "2015-01-23 10:00:00 PST" "2015-01-23 11:00:00 PST"
#>  [541] "2015-01-23 12:00:00 PST" "2015-01-23 13:00:00 PST"
#>  [543] "2015-01-23 14:00:00 PST" "2015-01-23 15:00:00 PST"
#>  [545] "2015-01-23 16:00:00 PST" "2015-01-23 17:00:00 PST"
#>  [547] "2015-01-23 18:00:00 PST" "2015-01-23 19:00:00 PST"
#>  [549] "2015-01-23 20:00:00 PST" "2015-01-23 21:00:00 PST"
#>  [551] "2015-01-23 22:00:00 PST" "2015-01-23 23:00:00 PST"
#>  [553] "2015-01-24 00:00:00 PST" "2015-01-24 01:00:00 PST"
#>  [555] "2015-01-24 02:00:00 PST" "2015-01-24 03:00:00 PST"
#>  [557] "2015-01-24 04:00:00 PST" "2015-01-24 05:00:00 PST"
#>  [559] "2015-01-24 06:00:00 PST" "2015-01-24 07:00:00 PST"
#>  [561] "2015-01-24 08:00:00 PST" "2015-01-24 09:00:00 PST"
#>  [563] "2015-01-24 10:00:00 PST" "2015-01-24 11:00:00 PST"
#>  [565] "2015-01-24 12:00:00 PST" "2015-01-24 13:00:00 PST"
#>  [567] "2015-01-24 14:00:00 PST" "2015-01-24 15:00:00 PST"
#>  [569] "2015-01-24 16:00:00 PST" "2015-01-24 17:00:00 PST"
#>  [571] "2015-01-24 18:00:00 PST" "2015-01-24 19:00:00 PST"
#>  [573] "2015-01-24 20:00:00 PST" "2015-01-24 21:00:00 PST"
#>  [575] "2015-01-24 22:00:00 PST" "2015-01-24 23:00:00 PST"
#>  [577] "2015-01-25 00:00:00 PST" "2015-01-25 01:00:00 PST"
#>  [579] "2015-01-25 02:00:00 PST" "2015-01-25 03:00:00 PST"
#>  [581] "2015-01-25 04:00:00 PST" "2015-01-25 05:00:00 PST"
#>  [583] "2015-01-25 06:00:00 PST" "2015-01-25 07:00:00 PST"
#>  [585] "2015-01-25 08:00:00 PST" "2015-01-25 09:00:00 PST"
#>  [587] "2015-01-25 10:00:00 PST" "2015-01-25 11:00:00 PST"
#>  [589] "2015-01-25 12:00:00 PST" "2015-01-25 13:00:00 PST"
#>  [591] "2015-01-25 14:00:00 PST" "2015-01-25 15:00:00 PST"
#>  [593] "2015-01-25 16:00:00 PST" "2015-01-25 17:00:00 PST"
#>  [595] "2015-01-25 18:00:00 PST" "2015-01-25 19:00:00 PST"
#>  [597] "2015-01-25 20:00:00 PST" "2015-01-25 21:00:00 PST"
#>  [599] "2015-01-25 22:00:00 PST" "2015-01-25 23:00:00 PST"
#>  [601] "2015-01-26 00:00:00 PST" "2015-01-26 01:00:00 PST"
#>  [603] "2015-01-26 02:00:00 PST" "2015-01-26 03:00:00 PST"
#>  [605] "2015-01-26 04:00:00 PST" "2015-01-26 05:00:00 PST"
#>  [607] "2015-01-26 06:00:00 PST" "2015-01-26 07:00:00 PST"
#>  [609] "2015-01-26 08:00:00 PST" "2015-01-26 09:00:00 PST"
#>  [611] "2015-01-26 10:00:00 PST" "2015-01-26 11:00:00 PST"
#>  [613] "2015-01-26 12:00:00 PST" "2015-01-26 13:00:00 PST"
#>  [615] "2015-01-26 14:00:00 PST" "2015-01-26 15:00:00 PST"
#>  [617] "2015-01-26 16:00:00 PST" "2015-01-26 17:00:00 PST"
#>  [619] "2015-01-26 18:00:00 PST" "2015-01-26 19:00:00 PST"
#>  [621] "2015-01-26 20:00:00 PST" "2015-01-26 21:00:00 PST"
#>  [623] "2015-01-26 22:00:00 PST" "2015-01-26 23:00:00 PST"
#>  [625] "2015-01-27 00:00:00 PST" "2015-01-27 01:00:00 PST"
#>  [627] "2015-01-27 02:00:00 PST" "2015-01-27 03:00:00 PST"
#>  [629] "2015-01-27 04:00:00 PST" "2015-01-27 05:00:00 PST"
#>  [631] "2015-01-27 06:00:00 PST" "2015-01-27 07:00:00 PST"
#>  [633] "2015-01-27 08:00:00 PST" "2015-01-27 09:00:00 PST"
#>  [635] "2015-01-27 10:00:00 PST" "2015-01-27 11:00:00 PST"
#>  [637] "2015-01-27 12:00:00 PST" "2015-01-27 13:00:00 PST"
#>  [639] "2015-01-27 14:00:00 PST" "2015-01-27 15:00:00 PST"
#>  [641] "2015-01-27 16:00:00 PST" "2015-01-27 17:00:00 PST"
#>  [643] "2015-01-27 18:00:00 PST" "2015-01-27 19:00:00 PST"
#>  [645] "2015-01-27 20:00:00 PST" "2015-01-27 21:00:00 PST"
#>  [647] "2015-01-27 22:00:00 PST" "2015-01-27 23:00:00 PST"
#>  [649] "2015-01-28 00:00:00 PST" "2015-01-28 01:00:00 PST"
#>  [651] "2015-01-28 02:00:00 PST" "2015-01-28 03:00:00 PST"
#>  [653] "2015-01-28 04:00:00 PST" "2015-01-28 05:00:00 PST"
#>  [655] "2015-01-28 06:00:00 PST" "2015-01-28 07:00:00 PST"
#>  [657] "2015-01-28 08:00:00 PST" "2015-01-28 09:00:00 PST"
#>  [659] "2015-01-28 10:00:00 PST" "2015-01-28 11:00:00 PST"
#>  [661] "2015-01-28 12:00:00 PST" "2015-01-28 13:00:00 PST"
#>  [663] "2015-01-28 14:00:00 PST" "2015-01-28 15:00:00 PST"
#>  [665] "2015-01-28 16:00:00 PST" "2015-01-28 17:00:00 PST"
#>  [667] "2015-01-28 18:00:00 PST" "2015-01-28 19:00:00 PST"
#>  [669] "2015-01-28 20:00:00 PST" "2015-01-28 21:00:00 PST"
#>  [671] "2015-01-28 22:00:00 PST" "2015-01-28 23:00:00 PST"
#>  [673] "2015-01-29 00:00:00 PST" "2015-01-29 01:00:00 PST"
#>  [675] "2015-01-29 02:00:00 PST" "2015-01-29 03:00:00 PST"
#>  [677] "2015-01-29 04:00:00 PST" "2015-01-29 05:00:00 PST"
#>  [679] "2015-01-29 06:00:00 PST" "2015-01-29 07:00:00 PST"
#>  [681] "2015-01-29 08:00:00 PST" "2015-01-29 09:00:00 PST"
#>  [683] "2015-01-29 10:00:00 PST" "2015-01-29 11:00:00 PST"
#>  [685] "2015-01-29 12:00:00 PST" "2015-01-29 13:00:00 PST"
#>  [687] "2015-01-29 14:00:00 PST" "2015-01-29 15:00:00 PST"
#>  [689] "2015-01-29 16:00:00 PST" "2015-01-29 17:00:00 PST"
#>  [691] "2015-01-29 18:00:00 PST" "2015-01-29 19:00:00 PST"
#>  [693] "2015-01-29 20:00:00 PST" "2015-01-29 21:00:00 PST"
#>  [695] "2015-01-29 22:00:00 PST" "2015-01-29 23:00:00 PST"
#>  [697] "2015-01-30 00:00:00 PST" "2015-01-30 01:00:00 PST"
#>  [699] "2015-01-30 02:00:00 PST" "2015-01-30 03:00:00 PST"
#>  [701] "2015-01-30 04:00:00 PST" "2015-01-30 05:00:00 PST"
#>  [703] "2015-01-30 06:00:00 PST" "2015-01-30 07:00:00 PST"
#>  [705] "2015-01-30 08:00:00 PST" "2015-01-30 09:00:00 PST"
#>  [707] "2015-01-30 10:00:00 PST" "2015-01-30 11:00:00 PST"
#>  [709] "2015-01-30 12:00:00 PST" "2015-01-30 13:00:00 PST"
#>  [711] "2015-01-30 14:00:00 PST" "2015-01-30 15:00:00 PST"
#>  [713] "2015-01-30 16:00:00 PST" "2015-01-30 17:00:00 PST"
#>  [715] "2015-01-30 18:00:00 PST" "2015-01-30 19:00:00 PST"
#>  [717] "2015-01-30 20:00:00 PST" "2015-01-30 21:00:00 PST"
#>  [719] "2015-01-30 22:00:00 PST" "2015-01-30 23:00:00 PST"
#>  [721] "2015-01-31 00:00:00 PST" "2015-01-31 01:00:00 PST"
#>  [723] "2015-01-31 02:00:00 PST" "2015-01-31 03:00:00 PST"
#>  [725] "2015-01-31 04:00:00 PST" "2015-01-31 05:00:00 PST"
#>  [727] "2015-01-31 06:00:00 PST" "2015-01-31 07:00:00 PST"
#>  [729] "2015-01-31 08:00:00 PST" "2015-01-31 09:00:00 PST"
#>  [731] "2015-01-31 10:00:00 PST" "2015-01-31 11:00:00 PST"
#>  [733] "2015-01-31 12:00:00 PST" "2015-01-31 13:00:00 PST"
#>  [735] "2015-01-31 14:00:00 PST" "2015-01-31 15:00:00 PST"
#>  [737] "2015-01-31 16:00:00 PST" "2015-01-31 17:00:00 PST"
#>  [739] "2015-01-31 18:00:00 PST" "2015-01-31 19:00:00 PST"
#>  [741] "2015-01-31 20:00:00 PST" "2015-01-31 21:00:00 PST"
#>  [743] "2015-01-31 22:00:00 PST" "2015-01-31 23:00:00 PST"
#>  [745] "2015-02-01 00:00:00 PST" "2015-02-01 01:00:00 PST"
#>  [747] "2015-02-01 02:00:00 PST" "2015-02-01 03:00:00 PST"
#>  [749] "2015-02-01 04:00:00 PST" "2015-02-01 05:00:00 PST"
#>  [751] "2015-02-01 06:00:00 PST" "2015-02-01 07:00:00 PST"
#>  [753] "2015-02-01 08:00:00 PST" "2015-02-01 09:00:00 PST"
#>  [755] "2015-02-01 10:00:00 PST" "2015-02-01 11:00:00 PST"
#>  [757] "2015-02-01 12:00:00 PST" "2015-02-01 13:00:00 PST"
#>  [759] "2015-02-01 14:00:00 PST" "2015-02-01 15:00:00 PST"
#>  [761] "2015-02-01 16:00:00 PST" "2015-02-01 17:00:00 PST"
#>  [763] "2015-02-01 18:00:00 PST" "2015-02-01 19:00:00 PST"
#>  [765] "2015-02-01 20:00:00 PST" "2015-02-01 21:00:00 PST"
#>  [767] "2015-02-01 22:00:00 PST" "2015-02-01 23:00:00 PST"
#>  [769] "2015-02-02 00:00:00 PST" "2015-02-02 01:00:00 PST"
#>  [771] "2015-02-02 02:00:00 PST" "2015-02-02 03:00:00 PST"
#>  [773] "2015-02-02 04:00:00 PST" "2015-02-02 05:00:00 PST"
#>  [775] "2015-02-02 06:00:00 PST" "2015-02-02 07:00:00 PST"
#>  [777] "2015-02-02 08:00:00 PST" "2015-02-02 09:00:00 PST"
#>  [779] "2015-02-02 10:00:00 PST" "2015-02-02 11:00:00 PST"
#>  [781] "2015-02-02 12:00:00 PST" "2015-02-02 13:00:00 PST"
#>  [783] "2015-02-02 14:00:00 PST" "2015-02-02 15:00:00 PST"
#>  [785] "2015-02-02 16:00:00 PST" "2015-02-02 17:00:00 PST"
#>  [787] "2015-02-02 18:00:00 PST" "2015-02-02 19:00:00 PST"
#>  [789] "2015-02-02 20:00:00 PST" "2015-02-02 21:00:00 PST"
#>  [791] "2015-02-02 22:00:00 PST" "2015-02-02 23:00:00 PST"
#>  [793] "2015-02-03 00:00:00 PST" "2015-02-03 01:00:00 PST"
#>  [795] "2015-02-03 02:00:00 PST" "2015-02-03 03:00:00 PST"
#>  [797] "2015-02-03 04:00:00 PST" "2015-02-03 05:00:00 PST"
#>  [799] "2015-02-03 06:00:00 PST" "2015-02-03 07:00:00 PST"
#>  [801] "2015-02-03 08:00:00 PST" "2015-02-03 09:00:00 PST"
#>  [803] "2015-02-03 10:00:00 PST" "2015-02-03 11:00:00 PST"
#>  [805] "2015-02-03 12:00:00 PST" "2015-02-03 13:00:00 PST"
#>  [807] "2015-02-03 14:00:00 PST" "2015-02-03 15:00:00 PST"
#>  [809] "2015-02-03 16:00:00 PST" "2015-02-03 17:00:00 PST"
#>  [811] "2015-02-03 18:00:00 PST" "2015-02-03 19:00:00 PST"
#>  [813] "2015-02-03 20:00:00 PST" "2015-02-03 21:00:00 PST"
#>  [815] "2015-02-03 22:00:00 PST" "2015-02-03 23:00:00 PST"
#>  [817] "2015-02-04 00:00:00 PST" "2015-02-04 01:00:00 PST"
#>  [819] "2015-02-04 02:00:00 PST" "2015-02-04 03:00:00 PST"
#>  [821] "2015-02-04 04:00:00 PST" "2015-02-04 05:00:00 PST"
#>  [823] "2015-02-04 06:00:00 PST" "2015-02-04 07:00:00 PST"
#>  [825] "2015-02-04 08:00:00 PST" "2015-02-04 09:00:00 PST"
#>  [827] "2015-02-04 10:00:00 PST" "2015-02-04 11:00:00 PST"
#>  [829] "2015-02-04 12:00:00 PST" "2015-02-04 13:00:00 PST"
#>  [831] "2015-02-04 14:00:00 PST" "2015-02-04 15:00:00 PST"
#>  [833] "2015-02-04 16:00:00 PST" "2015-02-04 17:00:00 PST"
#>  [835] "2015-02-04 18:00:00 PST" "2015-02-04 19:00:00 PST"
#>  [837] "2015-02-04 20:00:00 PST" "2015-02-04 21:00:00 PST"
#>  [839] "2015-02-04 22:00:00 PST" "2015-02-04 23:00:00 PST"
#>  [841] "2015-02-05 00:00:00 PST" "2015-02-05 01:00:00 PST"
#>  [843] "2015-02-05 02:00:00 PST" "2015-02-05 03:00:00 PST"
#>  [845] "2015-02-05 04:00:00 PST" "2015-02-05 05:00:00 PST"
#>  [847] "2015-02-05 06:00:00 PST" "2015-02-05 07:00:00 PST"
#>  [849] "2015-02-05 08:00:00 PST" "2015-02-05 09:00:00 PST"
#>  [851] "2015-02-05 10:00:00 PST" "2015-02-05 11:00:00 PST"
#>  [853] "2015-02-05 12:00:00 PST" "2015-02-05 13:00:00 PST"
#>  [855] "2015-02-05 14:00:00 PST" "2015-02-05 15:00:00 PST"
#>  [857] "2015-02-05 16:00:00 PST" "2015-02-05 17:00:00 PST"
#>  [859] "2015-02-05 18:00:00 PST" "2015-02-05 19:00:00 PST"
#>  [861] "2015-02-05 20:00:00 PST" "2015-02-05 21:00:00 PST"
#>  [863] "2015-02-05 22:00:00 PST" "2015-02-05 23:00:00 PST"
#>  [865] "2015-02-06 00:00:00 PST" "2015-02-06 01:00:00 PST"
#>  [867] "2015-02-06 02:00:00 PST" "2015-02-06 03:00:00 PST"
#>  [869] "2015-02-06 04:00:00 PST" "2015-02-06 05:00:00 PST"
#>  [871] "2015-02-06 06:00:00 PST" "2015-02-06 07:00:00 PST"
#>  [873] "2015-02-06 08:00:00 PST" "2015-02-06 09:00:00 PST"
#>  [875] "2015-02-06 10:00:00 PST" "2015-02-06 11:00:00 PST"
#>  [877] "2015-02-06 12:00:00 PST" "2015-02-06 13:00:00 PST"
#>  [879] "2015-02-06 14:00:00 PST" "2015-02-06 15:00:00 PST"
#>  [881] "2015-02-06 16:00:00 PST" "2015-02-06 17:00:00 PST"
#>  [883] "2015-02-06 18:00:00 PST" "2015-02-06 19:00:00 PST"
#>  [885] "2015-02-06 20:00:00 PST" "2015-02-06 21:00:00 PST"
#>  [887] "2015-02-06 22:00:00 PST" "2015-02-06 23:00:00 PST"
#>  [889] "2015-02-07 00:00:00 PST" "2015-02-07 01:00:00 PST"
#>  [891] "2015-02-07 02:00:00 PST" "2015-02-07 03:00:00 PST"
#>  [893] "2015-02-07 04:00:00 PST" "2015-02-07 05:00:00 PST"
#>  [895] "2015-02-07 06:00:00 PST" "2015-02-07 07:00:00 PST"
#>  [897] "2015-02-07 08:00:00 PST" "2015-02-07 09:00:00 PST"
#>  [899] "2015-02-07 10:00:00 PST" "2015-02-07 11:00:00 PST"
#>  [901] "2015-02-07 12:00:00 PST" "2015-02-07 13:00:00 PST"
#>  [903] "2015-02-07 14:00:00 PST" "2015-02-07 15:00:00 PST"
#>  [905] "2015-02-07 16:00:00 PST" "2015-02-07 17:00:00 PST"
#>  [907] "2015-02-07 18:00:00 PST" "2015-02-07 19:00:00 PST"
#>  [909] "2015-02-07 20:00:00 PST" "2015-02-07 21:00:00 PST"
#>  [911] "2015-02-07 22:00:00 PST" "2015-02-07 23:00:00 PST"
#>  [913] "2015-02-08 00:00:00 PST" "2015-02-08 01:00:00 PST"
#>  [915] "2015-02-08 02:00:00 PST" "2015-02-08 03:00:00 PST"
#>  [917] "2015-02-08 04:00:00 PST" "2015-02-08 05:00:00 PST"
#>  [919] "2015-02-08 06:00:00 PST" "2015-02-08 07:00:00 PST"
#>  [921] "2015-02-08 08:00:00 PST" "2015-02-08 09:00:00 PST"
#>  [923] "2015-02-08 10:00:00 PST" "2015-02-08 11:00:00 PST"
#>  [925] "2015-02-08 12:00:00 PST" "2015-02-08 13:00:00 PST"
#>  [927] "2015-02-08 14:00:00 PST" "2015-02-08 15:00:00 PST"
#>  [929] "2015-02-08 16:00:00 PST" "2015-02-08 17:00:00 PST"
#>  [931] "2015-02-08 18:00:00 PST" "2015-02-08 19:00:00 PST"
#>  [933] "2015-02-08 20:00:00 PST" "2015-02-08 21:00:00 PST"
#>  [935] "2015-02-08 22:00:00 PST" "2015-02-08 23:00:00 PST"
#>  [937] "2015-02-09 00:00:00 PST" "2015-02-09 01:00:00 PST"
#>  [939] "2015-02-09 02:00:00 PST" "2015-02-09 03:00:00 PST"
#>  [941] "2015-02-09 04:00:00 PST" "2015-02-09 05:00:00 PST"
#>  [943] "2015-02-09 06:00:00 PST" "2015-02-09 07:00:00 PST"
#>  [945] "2015-02-09 08:00:00 PST" "2015-02-09 09:00:00 PST"
#>  [947] "2015-02-09 10:00:00 PST" "2015-02-09 11:00:00 PST"
#>  [949] "2015-02-09 12:00:00 PST" "2015-02-09 13:00:00 PST"
#>  [951] "2015-02-09 14:00:00 PST" "2015-02-09 15:00:00 PST"
#>  [953] "2015-02-09 16:00:00 PST" "2015-02-09 17:00:00 PST"
#>  [955] "2015-02-09 18:00:00 PST" "2015-02-09 19:00:00 PST"
#>  [957] "2015-02-09 20:00:00 PST" "2015-02-09 21:00:00 PST"
#>  [959] "2015-02-09 22:00:00 PST" "2015-02-09 23:00:00 PST"
#>  [961] "2015-02-10 00:00:00 PST" "2015-02-10 01:00:00 PST"
#>  [963] "2015-02-10 02:00:00 PST" "2015-02-10 03:00:00 PST"
#>  [965] "2015-02-10 04:00:00 PST" "2015-02-10 05:00:00 PST"
#>  [967] "2015-02-10 06:00:00 PST" "2015-02-10 07:00:00 PST"
#>  [969] "2015-02-10 08:00:00 PST" "2015-02-10 09:00:00 PST"
#>  [971] "2015-02-10 10:00:00 PST" "2015-02-10 11:00:00 PST"
#>  [973] "2015-02-10 12:00:00 PST" "2015-02-10 13:00:00 PST"
#>  [975] "2015-02-10 14:00:00 PST" "2015-02-10 15:00:00 PST"
#>  [977] "2015-02-10 16:00:00 PST" "2015-02-10 17:00:00 PST"
#>  [979] "2015-02-10 18:00:00 PST" "2015-02-10 19:00:00 PST"
#>  [981] "2015-02-10 20:00:00 PST" "2015-02-10 21:00:00 PST"
#>  [983] "2015-02-10 22:00:00 PST" "2015-02-10 23:00:00 PST"
#>  [985] "2015-02-11 00:00:00 PST" "2015-02-11 01:00:00 PST"
#>  [987] "2015-02-11 02:00:00 PST" "2015-02-11 03:00:00 PST"
#>  [989] "2015-02-11 04:00:00 PST" "2015-02-11 05:00:00 PST"
#>  [991] "2015-02-11 06:00:00 PST" "2015-02-11 07:00:00 PST"
#>  [993] "2015-02-11 08:00:00 PST" "2015-02-11 09:00:00 PST"
#>  [995] "2015-02-11 10:00:00 PST" "2015-02-11 11:00:00 PST"
#>  [997] "2015-02-11 12:00:00 PST" "2015-02-11 13:00:00 PST"
#>  [999] "2015-02-11 14:00:00 PST" "2015-02-11 15:00:00 PST"
#>  [ reached getOption("max.print") -- omitted 7760 entries ]
#>  [ reached getOption("max.print") -- omitted 7760 entries ]